dom/media/tests/mochitest/test_peerConnection_bug822674.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    bug: "822674",
    title: "mozRTCPeerConnection isn't a true javascript object as it should be"
  });

  runTest(function () {
    var pc = new mozRTCPeerConnection();

    pc.thereIsNeverGoingToBeAPropertyWithThisNameOnThisInterface = 1;
    is(pc.thereIsNeverGoingToBeAPropertyWithThisNameOnThisInterface, 1,
       "Can set expandos on an RTCPeerConnection");

    pc = null;
    SimpleTest.finish();
  }, true);
</script>
</pre>
</body>
</html>
